# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GinBlog博客系统API接口文档",
        "title": "GinBlog博客系统",
        "contact": {},
        "version": "1.0.1"
    },
    "paths": {
        "/api/GetCategoryChildrenInfo": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取分类及子分类及每个分类下的已发布文章数量",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章分类相关接口"
                ],
                "summary": "获取分类及子分类",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
        "/api/GetProvinceAccess": {
            "get": {
                "description": "返回各省份的访问总量、最后访问时间，支持分页和排序",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "搜索统计相关接口"
                ],
                "summary": "查询各省份访问量统计",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码（默认1，最小1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数（默认20，最大100）",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段（access_count/last_access_at，默认access_count）",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方式（asc/desc，默认desc）",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/model.ProvincePageResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/model.ProvinceAccessStatisticResp"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
        "/api/GetSearch": {
            "get": {
                "description": "1. IP统计：仅访问单篇文章ID时触发；2. 标签统计：仅单搜索框搜标签/多字段传tag_names时触发；3. 浏览量：仅访问单篇ID时+1",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "搜索统计相关接口"
                ],
                "summary": "搜索文章（单关键词/多维度+分页+精准统计）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码（默认1，最小1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数（默认20，最大100）",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "单搜索框关键词（数字=ID，文本=标题/描述/内容/标签名）",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "文章ID（精确搜索，浏览量+1，IP统计触发）",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "文章标题（模糊搜索）",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "文章内容（模糊搜索）",
                        "name": "content",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "文章描述（模糊搜索）",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "标签名列表（多字段标签搜索，标签统计触发）",
                        "name": "tag_names",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "搜索成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.PageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误（含标签不存在）",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
...
        "/api/GetTagSearch": {
            "get": {
                "description": "关联标签表与标签搜索统计表，返回各标签的名称、搜索次数、最后搜索时间（未被搜索过的标签搜索量为0）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "搜索统计相关接口"
                ],
                "summary": "查询标签搜索量统计",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码（默认1，最小1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数（默认20，最大100）",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段（search_count/last_search_at，默认search_count）",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方式（asc/desc，默认desc）",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功\" // 这里必须加model.前缀",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/model.TagPageResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/model.TagSearchStatisticResp"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
        "/api/SendCode": {
            "post": {
                "description": "邮箱验证码接口(发送间隙60s)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "邮箱相关接口"
                ],
                "summary": "邮箱验证码接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "收件人邮箱地址",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.EmailDto"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/VerifyCode": {
            "post": {
                "description": "邮箱验证码校验接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "邮箱相关接口"
                ],
                "summary": "邮箱验证码校验接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "收件人邮箱地址",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "邮箱验证码",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.EmailCodeVo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/admin/CreateCategory": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "添加文章分类信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章分类相关接口"
                ],
                "summary": "添加文章分类信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分类名",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "缩略名",
                        "name": "slug",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "父类ID",
                        "name": "parent_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "分类描述",
                        "name": "description",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.CreateCategory"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/admin/CreateEmail": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "新增邮箱接口信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "邮箱相关接口"
                ],
                "summary": "新增邮箱接口信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮箱服务器",
                        "name": "host",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "邮箱用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "授权码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "邮箱端口",
                        "name": "port",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "发件人名称",
                        "name": "from_name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.CreateEmail"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/admin/CreateTag": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建标签名（管理员和用户均可创建，自动区分创建者类型）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "标签相关接口"
                ],
                "summary": "创建标签名",
                "parameters": [
                    {
                        "type": "string",
                        "description": "标签名",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功返回标签信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
...
        "/api/admin/DeleteCategory": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除文章分类信息（需先删除子分类和关联文章）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章分类相关接口"
                ],
                "summary": "删除文章分类信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分类ID",
                        "name": "cid",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.DeleteCategory"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/admin/GetCategoryChildrenInfo": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取分类及子分类及每个分类下的已发布文章数量",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章分类相关接口"
                ],
                "summary": "获取分类及子分类",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
...
        "/api/admin/getAdminInfo": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取当前登录管理员的信息（从登录态中获取username）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台相关接口"
                ],
                "summary": "获取管理员信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.EmailDto": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
...
        "model.ProvincePageResponse": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "地区统计列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ProvinceAccessStatisticResp"
                    }
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "total": {
                    "description": "总省份数",
                    "type": "integer"
                },
                "total_page": {
                    "description": "总页数",
                    "type": "integer"
                }
            }
        },
...
        "result.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}